Move video chapters into the sidebar lesson accordion
Instead of a chapter list under the video, the active lesson in the left
sidebar now expands to show its video chapters. Clicking a chapter seeks
the player, and the currently-playing chapter is highlighted.

- PlayerProvider: shares one player.js instance across the layout so the
  sidebar and the video (in separate subtrees) talk to the same player
- layout.js: wrap sidebar + content in PlayerProvider
- LessonVideo: mount iframe on the shared ref; publish chapters to the
  provider; drop the local chapter list
- CourseSidebar: render chapters as an accordion body under the active
  lesson, seek on click, highlight the active chapter
