+++ app/learn/[slug]/layout.js
@@ -1,5 +1,6 @@
 import { getCourseDetail } from "@/services/courses/getCourse";
 import CourseSidebar from "./CourseSidebar";
+import { PlayerProvider } from "./PlayerProvider";
 import Header from "@/components/header/header";
 import { notFound } from "next/navigation";
 
@@ -11,14 +12,16 @@ export default async function CourseLayout({ children, params }) {
   return (
     <div className="flex flex-col h-screen overflow-hidden">
       <Header />
-      <div className="flex flex-1 overflow-hidden">
-        <CourseSidebar course={course} courseSlug={slug} />
-        <main className="flex-1 overflow-y-auto">
-          <div className="mx-auto max-w-3xl px-16 py-10">
-            {children}
-          </div>
-        </main>
-      </div>
+      <PlayerProvider>
+        <div className="flex flex-1 overflow-hidden">
+          <CourseSidebar course={course} courseSlug={slug} />
+          <main className="flex-1 overflow-y-auto">
+            <div className="mx-auto max-w-3xl px-16 py-10">
+              {children}
+            </div>
+          </main>
+        </div>
+      </PlayerProvider>
     </div>
   );
 }
